% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bang.R
\name{bang}
\alias{bang}
\title{Causal discovery with latent confounding and non-Gaussian data}
\usage{
bang(Y, K, level = 0.01, verbose = T, restrict = 1)
}
\arguments{
\item{Y}{n x p matrix of observations with sample in row, variable in column}

\item{K}{The degree of the moment which is non-Gaussianity (i.e., this is K in the paper)}

\item{level}{the nominal level for the empirical likelihood tests}

\item{verbose}{Print progress updates}

\item{restrict}{Which moment restrictions to test in empirical likelihood
 \itemize{
 \item restrict = 1 restricts all moments up to degree K; i.e. gamma_v^k gamma_u and gamma_v gamma_u^k  for k < K
 \item restrict = 2, tests moments of degree K-1 and 1; i.e. gamma_v^(K-1) gamma_u and gamma_v gamma_u^(K-1)
}}
}
\value{
\itemize{
\item totalEffect estimate of (I-B)^(-1)
\item directEffect estimate of B
\item omega estimate of Omega
\item dEdge estimate of adjacncey matrix corresponding to directed edges
\item bEdge estimate of adjacncey matrix corresponding to bidirected edges
\item errs estimated epsilons
}
}
\description{
Estimates a graph from data generated by a linear SEM with non-Gaussian errors corresponding to a BAP
}
\examples{
\dontrun{
dat <- ngBap::rBAP(n = 50000, p = 7, dist = "gamma", d = 3, b = 5, ancestral = F, shuffle = T, signs = T)
Y <- dat$Y
out <- ngBap::bang(Y, K = 3, level = .01, verbose = F, restrict = 1)
}


}
