% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBAP.R
\name{rBAP}
\alias{rBAP}
\title{Causal discovery with latent confounding and non-Gaussian data}
\usage{
rBAP(
  n,
  p,
  dist,
  d,
  b,
  B = NULL,
  Omega = NULL,
  directLow = 0.6,
  directHigh = 1,
  bidirectLow = 0.3,
  bidirectHigh = 0.5,
  t.df = 13,
  ancestral = F,
  shuffle = T,
  signs = T
)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{dist}{the distribution of the errors
\itemize{
\item  "gauss" for normal distribution
\item "t" for T distribution with t.df degrees of freedom
\item "gamma" for gamma distribution
\item  "unif" for uniform distribution
\item "laplace" for Laplace distribution
\item "lognormal" for lognormal distribution
}}

\item{d}{number of directed edges}

\item{b}{number of bidirected edges}

\item{B}{the direct effects; leave as null to randomly sample}

\item{Omega}{the covariance of epsilon; leave as null to randomly sample}

\item{directLow}{the lower bound for the absolute value of a randomly sampled directed edge weight}

\item{directHigh}{the upper bound for the absolute value of a randomly sampled directed edge weight}

\item{bidirectLow}{the lower bound for the absolute value of a randomly sampled bidirected edge weight}

\item{bidirectHigh}{the upper bound for the absolute value of a randomly sampled bidirected edge weight}

\item{t.df}{degree of freedom if distribution is T}

\item{ancestral}{T restricts the graph to be ancestral; F allows for a BAP which may not be ancestral}

\item{shuffle}{Setting shuffle = T randomly relabels the nodes; if set to F then 1... p will always is a valid causal ordering}

\item{signs}{Setting signs = F, all parameters are positive; setting signs = F allows both positive and negative values}
}
\value{
\itemize{
\item B the directed edge coefficients
\item Omega covariance of varepsilon
\item directEdges the adjacency matrix of the directed edges (support of B)
\item bidirectEdgesthe adjacency matrix of the bidirected edges (support of Omega)
\item errs the sampled epsilon values
\item Y the sampled Y values
\item Sigma the population covariance
\item ord a topological ordering of the nodes
}
}
\description{
Samples from a linear SEM corresponding to a BAP
}
